feat : 관심매물 기능 추가
aptSeq로 조회해서 비슷한 매물 조회건 에러가 발생하였는데, 삭제하고 dealNo이라는 고유한 숫자를 사용해서 중복 조회를 방지함

+++ src/main/java/com/ssafy/zipcheck/interests/dto/InterestResponse.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class InterestResponse {
     // From interests table
-    private Long interestId;
+    private Integer interestId;
 
     // From houseinfos table
     private String aptSeq;
@@ -17,7 +17,7 @@ public class InterestResponse {
     private Double longitude;
 
     // From housedeals table
-    private Long no; // deal number
+    private Integer no; // deal number
     private String aptDong;
     private String floor;
     private Integer dealYear;
